Drop "reflect" package using proper type constraints
You don't need to use "reflect" package to check if pointers are nil/not nil if you use proper type constraints. You
can also use those to ensure that you are actually getting the pointer type, and don't check it in the runtime.

Drop unused variables and constants while we are at it.